Fixed crash by means of extra validation
The program should no longer freeze as it is no longer possible to update
the 3D view or create the shape without entering in all the details for a
component shape.
If any fields don't contain a number that's convertable to double, the
program will show an error message.
The error doesn't show what's wrong specifically, it just tells the user
there is something to check.

Refs #3737