Remove Workspace::setName method
The name is now a pure lookup into the ADS. The rename/remove methods
no longer need to be overloaded in the ADS so have been removed. The
add/addOrReplace are still required for the additional WorkspaceGroup
behaviour.
The add/addOrReplace methods have been refactored and a bug fixed where
the name on the workspace could be updated even if the DataService method
was going to throw.
Refs #6187