Some shapes confirmed to work. Complement identification works
At least spheres and finite cylinders import properly, the code is there for sliceofcylinderring and infinitecylinder but they've not been tested. The other shapes are awaiting parseing code.

I had made soem mistakes when i came to complements of shapes. The first was that even if they were idenitfied they weren't getting set os them, this was because i'd accidently redeclared the flag as local and it was going out of scope. the seocnd was that eh displayed text was worng due to the way i was constructing the QString.

Fixed a possible memeory leak (or at least objects kept in memeory and not deleted) as when i was parsing XML int he shapes some lists weren't being released.

Refs #3737