Tidied up some code after my attepts to find the cause
The cause of the Cone not rendering is that it's bounding box isn't calculated and it just returns the bounds given to it. So it is actually rendering, it's jsut so small not even a pixel is small enough to represent it.

As calculating this is beyond my skill i'm going to have to leave this ticket to someone else, so i've fixed the code i'd altered in my search for the problem and tidied some other areas.

Refs #7802