Refs #5012 Swap IDA interface to run LoadNexus C++ side.
This algorithm was being executed on a separate thread which was causing
the automatic file searching to play up in certain cases where it would
still be searching for files while the user had clicked the run button.

This change stops that from happening, but this should really be changed
to use AlgorithmRunner. But that's outside the scope of this ticket.