refs #9840. Issue encountered with multiperiod.
All seems to be running so far, but the PolarizationCorrection called from ReflectometryReductionOne is failing because it is not being passed the full input group workspace. This is most likely because ReflectometryReductionOne inherits from ReflectometryWorkflowBase, which inherits from DataProcessingAlgorithm, which inherits from Algorithm. I cannot inject the proper group workpsace processing algorithm (via MultiPeriodGroupAlgorithm) into this chain easily without making major structural changes.

The suggested way forwards would be to modify MultiPeriodGroupAlgorithm so that it's working were available as static methods. so that ReflectometryReductionOne could override ProcessGroups and CheckGroups and then point to the static methods in MultiPeriodGroupAlgorithm without technically inheriting from it.